Changing return type

+++ src/node_io.c
@@ -155,9 +155,8 @@ inline size_t cb_realfreespace(size_t free_space){
  * @param cb Circular buffer from where release the space
  * @param nbyte Number of bytes to release
  * @return 0 in case of success, -1 otherwise
- * TODO change return type
  */
-ssize_t cb_release(struct c_buf *cb, size_t nbyte){
+int cb_release(struct c_buf *cb, size_t nbyte){
 
     PTH_ERRCK_NC(pthread_mutex_lock(&cb->lock_refs))
 
@@ -174,9 +173,8 @@ ssize_t cb_release(struct c_buf *cb, size_t nbyte){
  * @param cb Circular buffer in object
  * @param nbyte Number of bytes to acquire
  * @return 0 in case of success, -1 otherwise
- * TODO change return type
  */
-ssize_t cb_acquire(struct c_buf *cb, size_t nbyte){
+int cb_acquire(struct c_buf *cb, size_t nbyte){
 
     PTH_ERRCK_NC(pthread_mutex_lock(&cb->lock_refs))
